(PC-12807) fix(EduConnectForm): Open the educonnect modal on mobile and desktop
